refactor: enhance entity
