+++ src/entity/aplication/usecases/find-by-id.usecase.ts
@@ -3,9 +3,7 @@ import ErrorEntityNotFound from '../../domain/errors/entity-not-found.error'
 import type EntityRepository from '../../domain/repositories/entity.repository'
 
 class FindEntityByIDUseCase {
-  constructor(private readonly repository: EntityRepository) {
-    this.repository = repository
-  }
+  constructor(private readonly repository: EntityRepository) {}
 
   async exec(id: string): Promise<EntityEntity> {
     const entityFound = await this.repository.find({ id })
