--- src/entity/domain/errors/entity-schema-validation.error.ts
@@ -1,8 +1,0 @@
-class ErrorEntitySchemaValidation extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = ErrorEntitySchemaValidation.name
-  }
-}
-
-export default ErrorEntitySchemaValidation
