+++ src/entity/domain/repositories/entity.repository.ts
@@ -2,7 +2,7 @@ import type Nullable from '../../../shared/domain/types/nullable.type'
 import type EntityEntity from '../entities/entity.entity'
 
 interface EntityRepository {
-  list: (offset: number, limit: number) => Promise<Nullable<EntityEntity[]>>
+  list: (offset?: number, limit?: number) => Promise<Nullable<EntityEntity[]>>
   find: (property: Record<string, string>) => Promise<Nullable<EntityEntity>>
   findWithPassword: (property: Record<string, string>) => Promise<Nullable<EntityEntity>>
   save: (entity: EntityEntity) => Promise<Nullable<EntityEntity>>
